Optic Flow Commit 7. Authored by Krister Blanch.
<Fixes>

<Features>
Added 4 basic tests.
    3 for the Radar Decoder.
    1 for the Testing library.

<New/Ongoing Bugs>
Same as prior commit.

<Compilability/Status>
Compiles with Docker.

+++ src/radar-decoder.hpp
@@ -6,6 +6,6 @@
 #include <string>
 #include <utility>
 
-uint8_t decode (opendlv::proxy::RadarDetectionReading msg, std::unique_ptr<cluon::SharedMemory> &shmArgb, uint16_t addBk[2048][512*2], bool verbose, uint16_t origin, uint16_t c_height, uint16_t c_width);
+int decode (opendlv::proxy::RadarDetectionReading msg, std::unique_ptr<cluon::SharedMemory> &shmArgb, uint16_t addBk[2048][512*2], bool verbose, uint16_t origin, uint16_t c_height, uint16_t c_width);
 
 #endif
